+++ sklearn/_build_utils/deprecated_modules.py
@@ -47,9 +47,9 @@ _DEPRECATED_MODULES = [
      'SpectralBiclustering'),
     ('_birch', 'sklearn.cluster.birch', 'sklearn.cluster', 'Birch'),
     ('_dbscan', 'sklearn.cluster.dbscan_', 'sklearn.cluster', 'DBSCAN'),
-    ('_hierarchical', 'sklearn.cluster.hierarchical', 'sklearn.cluster',
+    ('_hierarchical_impl', 'sklearn.cluster.hierarchical', 'sklearn.cluster',
      'FeatureAgglomeration'),
-    ('_k_means', 'sklearn.cluster.k_means_', 'sklearn.cluster', 'KMeans'),
+    ('_k_means_impl', 'sklearn.cluster.k_means_', 'sklearn.cluster', 'KMeans'),
     ('_mean_shift', 'sklearn.cluster.mean_shift_', 'sklearn.cluster',
      'MeanShift'),
     ('_optics', 'sklearn.cluster.optics_', 'sklearn.cluster', 'OPTICS'),
@@ -101,7 +101,7 @@ _DEPRECATED_MODULES = [
     ('_kernel_pca', 'sklearn.decomposition.kernel_pca',
      'sklearn.decomposition', 'KernelPCA'),
     ('_nmf', 'sklearn.decomposition.nmf', 'sklearn.decomposition', 'NMF'),
-    ('_online_lda', 'sklearn.decomposition.online_lda',
+    ('_online_lda_impl', 'sklearn.decomposition.online_lda',
      'sklearn.decomposition', 'LatentDirichletAllocation'),
     ('_online_lda_fast', 'sklearn.decomposition.online_lda_fast',
      'sklearn.decomposition', 'mean_change'),
@@ -140,7 +140,7 @@ _DEPRECATED_MODULES = [
 
     ('_dict_vectorizer', 'sklearn.feature_extraction.dict_vectorizer',
      'sklearn.feature_extraction', 'DictVectorizer'),
-    ('_hashing', 'sklearn.feature_extraction.hashing',
+    ('_hashing_impl', 'sklearn.feature_extraction.hashing',
      'sklearn.feature_extraction', 'FeatureHasher'),
     ('_stop_words', 'sklearn.feature_extraction.stop_words',
      'sklearn.feature_extraction.text', 'ENGLISH_STOP_WORDS'),
