+++ sklearn/cluster/_hierarchical_impl.py
