+++ sklearn/cluster/tests/test_hierarchical.py
@@ -22,8 +22,9 @@ from sklearn.utils._testing import ignore_warnings
 
 from sklearn.cluster import ward_tree
 from sklearn.cluster import AgglomerativeClustering, FeatureAgglomeration
-from sklearn.cluster._hierarchical import (_hc_cut, _TREE_BUILDERS,
-                                           linkage_tree, _fix_connectivity)
+from sklearn.cluster._hierarchical_impl import (_hc_cut, _TREE_BUILDERS,
+                                                linkage_tree,
+                                                _fix_connectivity)
 from sklearn.feature_extraction.image import grid_to_graph
 from sklearn.metrics.pairwise import PAIRED_DISTANCES, cosine_distances,\
     manhattan_distances, pairwise_distances
