+++ sklearn/decomposition/__init__.py
@@ -16,7 +16,7 @@ from ._dict_learning import (dict_learning, dict_learning_online,
                              MiniBatchDictionaryLearning, SparseCoder)
 from ._factor_analysis import FactorAnalysis
 from ..utils.extmath import randomized_svd
-from ._online_lda import LatentDirichletAllocation
+from ._online_lda_impl import LatentDirichletAllocation
 
 __all__ = ['DictionaryLearning',
            'FastICA',
