+++ sklearn/decomposition/_online_lda_impl.py
