+++ sklearn/feature_extraction/__init__.py
@@ -5,7 +5,7 @@ images.
 """
 
 from ._dict_vectorizer import DictVectorizer
-from ._hashing import FeatureHasher
+from ._hashing_impl import FeatureHasher
 from .image import img_to_graph, grid_to_graph
 from . import text
