+++ sklearn/feature_extraction/_hashing_impl.py
